move to armo types
